bytes: split bytes into separate compilation units
Makes for finer grained libraries in the bazel build.
